Fixed datafile issue.
The tests/gsa contains two datafiles with the same basename and different extensions (.m and .mat). The new behaviour of the data interface forces the user to choose between the two candidate data files (ie it is no more possible to load a data file without specifying the extension in this case).