Fixed bug (appearing only with Matlab and legacy data interface) if the m data file is in the current directory.
If the m data file is in the current directory, the first output of
fileparts(options_.datafile) is an empty string. In Matlab the command
cd('') fails, but Octave does not fail in this case (and does nothing).

Added a test on the first output of fileparts, if this output is empty
we do not try to change the current directory.