Fix length of stored and plotted forecasts for Bayesian estimation
It stored the initial condition(s) in addition to the forecasts for the specified horizon. In plots, the first point was actually not the forecast, but the initial condition. This contrasted with the behavior or ML and recursive Bayesian forecasts. Now the stored and plotted forecasts always have the first forecast horizon as the first entry/data point