Replaced get_prior_info routine by prior routine.
This new routine gives the possibility to obtain a description of the
prior from the command line.

>> prior table

will display a table with the prior mean, prior mode, prior std, prior
bounds and prior HPD interval.

>> prior plot

will plot the prior densities.

>> prior optimize

will trigger the maximization of the prior density and display the
results (often, because of BK conditions or other issues, Dynare is even
not able to get the prior mode).

>> prior moments

will display the moments of the endogenous variables at the prior mode.

>> prior simulate

will run a Monte-Carlo, by sampling from the prior and discarding vector
of parameters such that the steady state does not exist or the BK
conditions are not met, and return an estimate of the effective prior
mass.