Check that Dynare's routines are not overridden by other routines.
By default Dynare will temporarly change Matlab/Octave's path. Added an
option to force Dynare to keep the current path. Also check that the user
did not put dynare/matlab directory with all the subfolders in the path.

With this patch, we do not need to prefix Dynare routines with dyn_ or
dynare_ anymore.