Fixed ep.init option.
Initialization of the perfect foresight solver (in extended path) with
the solution of the first order approximation of the model was broken.

If the value of options_.ep.init is "true"" (1) then the solution of the
first order approximation is used as an initial guess for the newton
lilke solver. If the value of options_.ep.init is "false" (0) the solver
is initialized with the steady state.