Removed one (failing) integration test.
fs2000/fs2000_analytical_derivation.mod was almost a duplicate of
analytical_derivates/fs2000_analytical_derivation.mod.

The removed integration test was failing because of the sign of the
hessian matrix. The only significant difference between the two mod
files, was that the removed one did not use the closed form expression
of the steady state.